main_process/requirements.txt
numpy == 1.21.2
pandas == 1.3.3
scikit-learn == 0.24.2
google-cloud-storage==1.42.2
google-api-python-client==2.20.0